+++ src/featureInfo/balloonHelper.js
@@ -2,15 +2,26 @@ import {
   Cartesian2,
   Cartographic,
   SceneTransforms,
+  HeightReference,
+  Entity,
 } from '@vcmap-cesium/engine';
 import {
+  cartesianToMercator,
   CesiumMap,
+  getGeometryHeight,
+  getHeightInfo,
+  isAbsoluteHeightReference,
+  isRelativeHeightReference,
   ObliqueMap,
   OpenlayersMap,
   Projection,
   transformToImage,
+  VectorProperties,
 } from '@vcmap/core';
 import { unByKey } from 'ol/Observable.js';
+import Feature from 'ol/Feature.js';
+import { Point } from 'ol/geom.js';
+import { getCenter } from 'ol/extent.js';
 import {
   getWindowPositionOptionsFromMapEvent,
   WindowAlignment,
@@ -172,4 +183,70 @@ export async function setupBalloonPositionListener(
   await setup(vcsApp, windowId, clickedPosition);
 
   return destroy;
+}
+
+/**
+ * @param {Feature} feature
+ * @param {import("@vcmap/core").Layer} layer
+ * @param {import("ol/coordinate.js").Coordinate|undefined} clickedPosition
+ * @returns {{position?: import("ol/coordinate.js").Coordinate, heightReference: HeightReference, heightOffset: number}}
+ */
+export function getBalloonPositionFromFeature(feature, layer, clickedPosition) {
+  let heightReference = HeightReference.NONE;
+  let heightOffset = 0;
+  let position = null;
+  if (feature instanceof Feature && feature.getGeometry() instanceof Point) {
+    const point = feature.getGeometry();
+    const vectorProperties =
+      layer.vectorProperties ??
+      layer.featureProvider?.vectorProperties ??
+      new VectorProperties();
+    const renderAs = vectorProperties.renderAs(feature);
+    if (renderAs === 'geometry') {
+      // special case where we do not want to use the clickedPosition but the exact Position of the Point
+      const heightInfo = getHeightInfo(feature, point, vectorProperties);
+      ({ heightReference } = heightInfo);
+      let height = clickedPosition?.[2] ?? 0;
+      position = point.getCoordinates();
+      // if clamped, do nothing
+      if (isRelativeHeightReference(heightReference)) {
+        height = getGeometryHeight(point, heightInfo);
+        if (heightInfo.groundLevel != null) {
+          // we have a groundLevel, so no need to clamp the point
+          heightReference = HeightReference.NONE;
+        }
+        if (heightInfo.heightAboveGround != null) {
+          heightOffset += heightInfo.heightAboveGround;
+        }
+        const extrudedHeight = heightInfo.storeyHeightsAboveGround.reduce(
+          (acc, storeyHeight) => acc + storeyHeight,
+          0,
+        );
+        heightOffset += extrudedHeight;
+        height += extrudedHeight;
+        position = [position[0], position[1], height];
+      } else if (isAbsoluteHeightReference(heightReference)) {
+        const extrudedHeight = heightInfo.storeyHeightsAboveGround.reduce(
+          (acc, storeyHeight) => acc + storeyHeight,
+          0,
+        );
+        position = [
+          position[0],
+          position[1],
+          heightInfo.groundLevelOrMinHeight + extrudedHeight,
+        ];
+      }
+      return { position, heightOffset, heightReference };
+    }
+  }
+  if (clickedPosition) {
+    position = clickedPosition.slice();
+  } else if (feature instanceof Feature && feature.getGeometry()) {
+    position = getCenter(feature.getGeometry().getExtent());
+  } else if (feature instanceof Entity) {
+    position = cartesianToMercator(feature.position);
+  } else if (feature?.primitive?.boundingSphere?.center) {
+    position = cartesianToMercator(feature.primitive.boundingSphere.center);
+  }
+  return { position, heightReference, heightOffset };
 }